+++ modules/butler.js
@@ -109,6 +109,10 @@ export const KNOWN_GAME_SYSTEMS = {
         healthAttribute: "",
         initiative: ""
     },
+    cpredcore: {
+        id: "cyberpunk-red-core",
+        name: "Cyberpunk Red Core"
+    },
     other: {
         id: "other",
         name: "Custom/Other",
@@ -199,6 +203,10 @@ export const DEFAULT_CONFIG = {
             id: "cub-enhanced-condition-trigger-config",
             title: "CUB Enhanced Condition - Trigger Config"
         },
+        optionConfig: {
+            id: "cub-enhanced-condition-option-config",
+            title: "CUB Enhanced Condition - Option Config"
+        },
         title: "Enhanced Conditions",
         mapTypes: {
             default: "System - Default",
@@ -233,9 +241,18 @@ export const DEFAULT_CONFIG = {
             chatConditionsPartial: `${PATH}/templates/partials/chat-card-condition-list.hbs`,
             importDialog: `${PATH}/templates/import-conditions.html`,
             macroConfig: `${PATH}/templates/enhanced-condition-macro-config.hbs`,
-            triggerConfig: `${PATH}/templates/enhanced-condition-trigger-config.hbs`
+            triggerConfig: `${PATH}/templates/enhanced-condition-trigger-config.hbs`,
+            optionConfig: `${PATH}/templates/enhanced-condition-option-config.hbs`
         },
-        migrationVersion: ""
+        migrationVersion: "",
+        specialStatusEffects: {
+            blinded: {
+                optionProperty: "blindToken"
+            },
+            invisible: {
+                optionProperty: "markInvisible"
+            }
+        }
     },
     giveXP: {
         enable: false,
@@ -429,7 +446,9 @@ export const SETTING_KEYS = {
         outputCombat: "conditionsOutputDuringCombat",
         suppressPreventativeSaveReminder: "conditionsSuppressPreventativeSaveReminder",
         migrationVersion: "enhancedConditionsMigrationVersion",
-        showSortDirectionDialog: "showSortDirectionDialog"
+        showSortDirectionDialog: "showSortDirectionDialog",
+        defaultSpecialStatusEffects: "defaultSpecialStatusEffects",
+        specialStatusEffectMapping: "specialStatusEffectMapping"
     },
     giveXP: {
         enable: "enableGiveXP",